Вынести хранение данных о фильмах и юзерах в отдельные классы InMemoryFilmStorage и InMemoryUserStorage
