+++ src/test/java/ru/yandex/practicum/filmorate/controller/UserControllerTest.java
@@ -6,6 +6,7 @@ import ru.yandex.practicum.filmorate.exception.DuplicateEmailException;
 import ru.yandex.practicum.filmorate.exception.DuplicateLoginException;
 import ru.yandex.practicum.filmorate.exception.UserNotFoundException;
 import ru.yandex.practicum.filmorate.model.User;
+import ru.yandex.practicum.filmorate.storage.user.InMemoryUserStorage;
 
 import java.time.LocalDate;
 
@@ -17,7 +18,7 @@ class UserControllerTest {
 
     @BeforeEach
     public void setUp() {
-        userController = new UserController();
+        userController = new UserController(new InMemoryUserStorage());
     }
 
     @Test
